+++ CHANGELOG.md
@@ -7,6 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2019-09-13
+
+### Added
+- New Page methods Page::getLayout, Page::setLayout(), Page::getView(), and Page::setView().
+- New Controller::render() method, executed right before a page is actually rendered.
+- New ViewData class for storing (internal) data required by the View class.
+- New getter/setter methods for ViewData properties for the View class.
+- New method Wireframe::getConfig() for getting current config settings.
+- New method ViewPlaceholders::has() for checking if a placeholder has already been populated.
+
+### Changed
+- Various View-related features moved from Wireframe module and ViewPlaceholders class to the View class.
+- Removed access to local get* and set* methods via the PHP's magic setter method __set() and getter method __get() in the View class.
+- Redirect feature no longer fails if provided with a WireArray data type; in these cases the first item is used as the redirect target.
+- Improvements to PHPDoc comments.
+
+### Fixed
+- An issue with Config class where the "all directories exist" message was sometimes displayed unintentionally.
+- An issue where View Placeholder values might've been overwritten because existence of earlier value was checked inproperly.
+- An issue where empty / null view file would be automatically replaced with value "default".
+
 ## [0.5.2] - 2019-08-28
 
 ### Fixed